Implemented AccountsController and fixed some controller and service classes logic.

+++ fintrack/src/main/java/com/studying/fintrack/domain/services/AccountsService.java
@@ -3,6 +3,8 @@ package com.studying.fintrack.domain.services;
 import com.studying.fintrack.domain.entities.Account;
 import com.studying.fintrack.domain.models.AccountDTO;
 import com.studying.fintrack.domain.repositories.AccountsRepository;
+import jakarta.persistence.EntityNotFoundException;
+import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -23,23 +25,37 @@ public class AccountsService {
     return accountsRepository.save(account);
   }
 
-  public Account getAccountById(Integer id) {
+  public Account getAccountById(int id) {
     Account account = accountsRepository.findById(id).orElse(null);
     if (account == null) {
-      throw new NullPointerException("Account not found in DB!");
+      throw new EntityNotFoundException("Account not found in DB!");
     }
     return account;
   }
 
-  public Account updateAccount(Account account) {
+  public Account updateAccount(int id, AccountDTO dto) {
+    Account account = accountsRepository.findById(id).orElse(null);
+    if (account != null) {
+      throw new EntityNotFoundException("Account not found in DB!");
+    }
+    account.setName(dto.getName());
+    account.setCurrency(dto.getCurrency());
     return accountsRepository.save(account);
   }
 
-  public void deleteAccount(Account account) {
+  public void deleteAccount(int id) {
+    Account account = accountsRepository.findById(id).orElse(null);
+    if (account == null) {
+      throw new EntityNotFoundException("Account not found in DB!");
+    }
     accountsRepository.delete(account);
   }
 
-  public Iterable<Account> getAllAccounts() {
+  public List<Account> getAllAccounts() {
     return accountsRepository.findAll();
+  }
+
+  public List<Account> getAccountsByUserId(int userId) {
+    return accountsRepository.findByUserId(userId);
   }
 }
